(MAINT) Homepage revamp
This commit adds a flexbox of cards to the homepage which display recently
published blog posts. It is also possible to pin one blog post to this
section using the `[params.home].pinnedPost` variable which has been added to the
site's `config.toml`.

Furthermore, a new heading and subheading has been added to the homepage and
the main contents of the page has been changed.